Split MultiplayerReadyButton
