+++ osu.Game/Screens/OnlinePlay/Multiplayer/Match/MatchStartControl.cs
@@ -4,32 +4,21 @@
 using System;
 using System.Diagnostics;
 using System.Linq;
-using Humanizer;
 using JetBrains.Annotations;
 using osu.Framework.Allocation;
 using osu.Framework.Audio;
 using osu.Framework.Audio.Sample;
 using osu.Framework.Bindables;
-using osu.Framework.Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Cursor;
-using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Localisation;
 using osu.Framework.Threading;
-using osu.Game.Graphics;
-using osu.Game.Graphics.Backgrounds;
-using osu.Game.Graphics.UserInterface;
-using osu.Game.Graphics.UserInterfaceV2;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.Multiplayer.Countdown;
 using osuTK;
 
 namespace osu.Game.Screens.OnlinePlay.Multiplayer.Match
 {
-    public class MultiplayerReadyButton : MultiplayerRoomComposite
+    public class MatchStartControl : MultiplayerRoomComposite
     {
         [Resolved]
         private OngoingOperationTracker ongoingOperationTracker { get; set; }
@@ -47,7 +36,7 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer.Match
         private ScheduledDelegate readySampleDelegate;
         private IBindable<bool> operationInProgress;
 
-        public MultiplayerReadyButton()
+        public MatchStartControl()
         {
             InternalChild = new GridContainer
             {
@@ -230,221 +219,6 @@ namespace osu.Game.Screens.OnlinePlay.Multiplayer.Match
 
                 countReady = newCountReady;
             });
-        }
-
-        public class ReadyButton : Components.ReadyButton
-        {
-            public new Triangles Triangles => base.Triangles;
-
-            [Resolved]
-            private MultiplayerClient multiplayerClient { get; set; }
-
-            [Resolved]
-            private OsuColour colours { get; set; }
-
-            [CanBeNull]
-            private MultiplayerRoom room => multiplayerClient.Room;
-
-            protected override void LoadComplete()
-            {
-                base.LoadComplete();
-
-                multiplayerClient.RoomUpdated += () => Scheduler.AddOnce(onRoomUpdated);
-                onRoomUpdated();
-            }
-
-            protected override void Update()
-            {
-                base.Update();
-
-                if (room?.Countdown != null)
-                {
-                    // Update the countdown timer.
-                    onRoomUpdated();
-                }
-            }
-
-            private void onRoomUpdated()
-            {
-                updateButtonText();
-                updateButtonColour();
-            }
-
-            private void updateButtonText()
-            {
-                if (room == null)
-                {
-                    Text = "Ready";
-                    return;
-                }
-
-                var localUser = multiplayerClient.LocalUser;
-
-                int countReady = room.Users.Count(u => u.State == MultiplayerUserState.Ready);
-                int countTotal = room.Users.Count(u => u.State != MultiplayerUserState.Spectating);
-                string countText = $"({countReady} / {countTotal} ready)";
-
-                if (room.Countdown != null)
-                {
-                    string countdownText = $"Starting in {room.Countdown.EndTime - DateTimeOffset.Now:mm\\:ss}";
-
-                    switch (localUser?.State)
-                    {
-                        default:
-                            Text = $"Ready ({countdownText.ToLowerInvariant()})";
-                            break;
-
-                        case MultiplayerUserState.Spectating:
-                        case MultiplayerUserState.Ready:
-                            Text = $"{countdownText} {countText}";
-                            break;
-                    }
-                }
-                else
-                {
-                    switch (localUser?.State)
-                    {
-                        default:
-                            Text = "Ready";
-                            break;
-
-                        case MultiplayerUserState.Spectating:
-                        case MultiplayerUserState.Ready:
-                            Text = room.Host?.Equals(localUser) == true
-                                ? $"Start match {countText}"
-                                : $"Waiting for host... {countText}";
-
-                            break;
-                    }
-                }
-            }
-
-            private void updateButtonColour()
-            {
-                if (room == null)
-                {
-                    setGreen();
-                    return;
-                }
-
-                var localUser = multiplayerClient.LocalUser;
-
-                switch (localUser?.State)
-                {
-                    default:
-                        setGreen();
-                        break;
-
-                    case MultiplayerUserState.Spectating:
-                    case MultiplayerUserState.Ready:
-                        if (room?.Host?.Equals(localUser) == true && room.Countdown == null)
-                            setGreen();
-                        else
-                            setYellow();
-
-                        break;
-                }
-
-                void setYellow()
-                {
-                    BackgroundColour = colours.YellowDark;
-                    Triangles.ColourDark = colours.YellowDark;
-                    Triangles.ColourLight = colours.Yellow;
-                }
-
-                void setGreen()
-                {
-                    BackgroundColour = colours.Green;
-                    Triangles.ColourDark = colours.Green;
-                    Triangles.ColourLight = colours.GreenLight;
-                }
-            }
-
-            protected override void Dispose(bool isDisposing)
-            {
-                base.Dispose(isDisposing);
-
-                if (multiplayerClient != null)
-                    multiplayerClient.RoomUpdated -= onRoomUpdated;
-            }
-
-            public override LocalisableString TooltipText
-            {
-                get
-                {
-                    if (room?.Countdown != null && multiplayerClient.IsHost && multiplayerClient.LocalUser?.State == MultiplayerUserState.Ready)
-                        return "Cancel countdown";
-
-                    return base.TooltipText;
-                }
-            }
-        }
-
-        public class CountdownButton : IconButton, IHasPopover
-        {
-            private static readonly TimeSpan[] available_delays =
-            {
-                TimeSpan.FromSeconds(10),
-                TimeSpan.FromSeconds(30),
-                TimeSpan.FromMinutes(1),
-                TimeSpan.FromMinutes(2)
-            };
-
-            public new Action<TimeSpan> Action;
-
-            private readonly Drawable background;
-
-            public CountdownButton()
-            {
-                Icon = FontAwesome.Solid.CaretDown;
-                IconScale = new Vector2(0.6f);
-
-                Add(background = new Box
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Depth = float.MaxValue
-                });
-
-                base.Action = this.ShowPopover;
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OsuColour colours)
-            {
-                background.Colour = colours.Green;
-            }
-
-            public Popover GetPopover()
-            {
-                var flow = new FillFlowContainer
-                {
-                    Width = 200,
-                    AutoSizeAxes = Axes.Y,
-                    Direction = FillDirection.Vertical,
-                    Spacing = new Vector2(2),
-                };
-
-                foreach (var duration in available_delays)
-                {
-                    flow.Add(new PopoverButton
-                    {
-                        RelativeSizeAxes = Axes.X,
-                        Text = $"Start match in {duration.Humanize()}",
-                        BackgroundColour = background.Colour,
-                        Action = () =>
-                        {
-                            Action(duration);
-                            this.HidePopover();
-                        }
-                    });
-                }
-
-                return new OsuPopover { Child = flow };
-            }
-
-            public class PopoverButton : OsuButton
-            {
-            }
         }
     }
 }
